Refactor overload methods into Functional class
